+++ recipes/update_head_node.rb
@@ -2,7 +2,7 @@
 
 #
 # Cookbook Name:: aws-parallelcluster
-# Recipe:: _undo_master_base_config
+# Recipe:: update_head_node
 #
 # Copyright 2013-2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
@@ -14,3 +14,8 @@
 # or in the "LICENSE.txt" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
 # OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
+
+
+if node['cfncluster']['cfn_scheduler'] == 'slurm'
+  include_recipe 'aws-parallelcluster::update_head_node_slurm'
+end