+++ CHANGELOG.md
@@ -3,6 +3,34 @@ aws-parallelcluster-cookbook CHANGELOG
 
 This file is used to list changes made in each version of the AWS ParallelCluster cookbook.
 
+2.10.1
+------
+
+**ENHANCEMENTS**
+- Install Arm Performance Libraries (APL) 20.2.1 on ARM AMIs (CentOS8, Alinux2, Ubuntu1804).
+- Install EFA kernel module on ARM instances with `alinux2` and `ubuntu1804`.
+- Configure NFS threads to be max(8, num_cores) for performance. This enhancement will not take effect on Ubuntu 16.04.
+
+**CHANGES**
+- Upgrade EFA installer to version 1.11.1.
+  - EFA configuration: ``efa-config-1.7`` (from efa-config-1.5)
+  - EFA profile: ``efa-profile-1.3`` (from efa-profile-1.1)
+  - EFA kernel module: ``efa-1.10.2`` (no change)
+  - RDMA core: ``rdma-core-31.2amzn`` (from rdma-core-31.amzn0)
+  - Libfabric: ``libfabric-1.11.1amzn1.0`` (from libfabric-1.11.1amzn1.1)
+  - Open MPI: ``openmpi40-aws-4.1.0`` (from openmpi40-aws-4.0.5)
+- Upgrade Intel MPI to version U8.
+- Upgrade NICE DCV to version 2020.2-9662.
+- Set default systemd runlevel to multi-user.target on all OSes during ParallelCluster official ami creation.
+  The runlevel is set to graphical.target on head node only when DCV is enabled. This prevents the execution of
+  graphical services, such as x/gdm, when they are not required.
+- Download Intel MPI and HPC packages from S3 rather than Intel yum repos.
+
+**BUG FIXES**
+- Fix installation of Intel PSXE package on CentOS 7 by using yum4.
+- Fix routing issues with multiple Network Interfaces on Ubuntu 18.04.
+- Fix compilation of SGE by downloading sources from Debian repository and not from the EOL Ubuntu 19.10.
+
 2.10.0
 ------